+++ mopidy/mpd/exceptions.py
@@ -93,24 +93,24 @@ class MpdNotImplemented(MpdAckError):
 
 
 class MpdInvalidTrackForPlaylist(MpdAckError):
+    # NOTE: This is a custom error for Mopidy that does not exist in MPD.
     error_code = 0
 
-    def __init__(self, backend_scheme, track_scheme, *args, **kwargs):
+    def __init__(self, playlist_scheme, track_scheme, *args, **kwargs):
         super(MpdInvalidTrackForPlaylist, self).__init__(*args, **kwargs)
-        self.message = 'Playlist backend "%s" can\'t store ' \
-            'track scheme "%s"' % (backend_scheme, track_scheme)
+        self.message = (
+            'Playlist with scheme "%s" can\'t store track scheme "%s"' %
+            (playlist_scheme, track_scheme))
 
 
 class MpdFailedToSavePlaylist(MpdAckError):
+    # NOTE: This is a custom error for Mopidy that does not exist in MPD.
     error_code = 0
 
     def __init__(self, backend_scheme, *args, **kwargs):
         super(MpdFailedToSavePlaylist, self).__init__(*args, **kwargs)
-        if backend_scheme is None:
-            self.message = 'Failed to save playlist'
-        else:
-            self.message = 'Backend "%s" failed to save playlist' % (
-                backend_scheme, )
+        self.message = 'Backend with scheme "%s" failed to save playlist' % (
+            backend_scheme)
 
 
 class MpdDisabled(MpdAckError):